refactor(payment): PI-147 PaymentMethodTitle moved spec to RTL
