+++ jest-setup.ts
@@ -1,5 +1,6 @@
 import '@testing-library/jest-dom';
 import '@testing-library/jest-dom/extend-expect';
+import { configure as configureRTL } from '@testing-library/react';
 import * as Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import { configure } from 'enzyme';
 import { noop } from 'lodash';
@@ -7,6 +8,7 @@ import { noop } from 'lodash';
 const adapter = Adapter as any;
 
 configure({ adapter: new adapter.default() });
+configureRTL({ testIdAttribute: 'data-test' });
 
 // https://github.com/facebook/jest/issues/10784
 process.on('unhandledRejection', (reason) => {
